Find all paths that add up to the sum
Given a Tree and a target number.
Come up with an algorithm that prints all paths which sum to a the
target number.
In this example, the path must start at the root.
To make it more fun, I will try to do it in a way that doesn’t require
the path to start at the root.